Undoing change to PulseGen - input field is back to integer. Some commented out debug prints in the other classes.

+++ device/PulseGen.h
@@ -115,8 +115,8 @@ class PulseGen
     int trigMode_;
     bool secondPulse_;
     
-    double prevInput_;
-    double input_;    
+    int prevInput_;
+    int input_;    
 };
